refactoring: move service/MethodsService to webservice/MethodsApiService
