+++ src/main/scala/org/broadinstitute/dsde/firecloud/webservice/MethodsApiService.scala
@@ -1,22 +1,23 @@
-package org.broadinstitute.dsde.firecloud.service
+package org.broadinstitute.dsde.firecloud.webservice
 
 import akka.actor.{Actor, Props}
 import org.broadinstitute.dsde.firecloud.FireCloudConfig
 import org.broadinstitute.dsde.firecloud.core.{AgoraPermissionActor, AgoraPermissionHandler}
 import org.broadinstitute.dsde.firecloud.model.MethodRepository._
 import org.broadinstitute.dsde.firecloud.model.ModelJsonProtocol._
+import org.broadinstitute.dsde.firecloud.service.{FireCloudDirectives, PerRequestCreator}
 import org.slf4j.LoggerFactory
 import spray.http.StatusCodes
 import spray.httpx.SprayJsonSupport._
 import spray.json.DefaultJsonProtocol._
 import spray.routing.{HttpService, Route}
 
-class MethodsServiceActor extends Actor with MethodsService {
+class MethodsApiServiceActor extends Actor with MethodsApiService {
   def actorRefFactory = context
   def receive = runRoute(routes)
 }
 
-object MethodsService {
+object MethodsApiService {
   val remoteMethodsPath = FireCloudConfig.Agora.authPrefix + "/methods"
   val remoteMethodsUrl = FireCloudConfig.Agora.baseUrl + remoteMethodsPath
   val remoteConfigurationsPath = FireCloudConfig.Agora.authPrefix + "/configurations"
@@ -24,7 +25,7 @@ object MethodsService {
   val remotePermissionsTemplate = FireCloudConfig.Agora.baseUrl + FireCloudConfig.Agora.authPrefix + "/%s/%s/%s/%s/permissions"
 }
 
-trait MethodsService extends HttpService with PerRequestCreator with FireCloudDirectives {
+trait MethodsApiService extends HttpService with PerRequestCreator with FireCloudDirectives {
 
   lazy val log = LoggerFactory.getLogger(getClass)
 
@@ -71,14 +72,14 @@ trait MethodsService extends HttpService with PerRequestCreator with FireCloudDi
   // Agora routes that can be passthroughs. Because these routes conflict with the override routes, make sure
   // they are processed second!
   val passthroughRoutes: Route =
-    passthroughAllPaths(localMethodsPath, MethodsService.remoteMethodsUrl) ~
-      passthroughAllPaths(localConfigsPath, MethodsService.remoteConfigurationsUrl)
+    passthroughAllPaths(localMethodsPath, MethodsApiService.remoteMethodsUrl) ~
+      passthroughAllPaths(localConfigsPath, MethodsApiService.remoteConfigurationsUrl)
 
   // combine all of the above route definitions, keeping overrides first.
   val routes =  methodsAndConfigsACLOverrideRoute ~ passthroughRoutes
 
 
   private def getUrlFromBasePath(agoraEntity: String, namespace: String, name: String, snapshotId: Int): String = {
-    MethodsService.remotePermissionsTemplate.format(agoraEntity, namespace, name, snapshotId)
+    MethodsApiService.remotePermissionsTemplate.format(agoraEntity, namespace, name, snapshotId)
   }
 }
